implement L2DataPublisher

+++ include/matching/L2Data.hpp
@@ -7,5 +7,5 @@ struct alignas(64) L2Data {
     int32_t quantity_;
     OrderSide side_;
 
-    int8_t sequence_id_;
+    int64_t sequence_id_;
 };
